+++ featuretools/primitives/api.py
@@ -1,5 +1,4 @@
 # flake8: noqa
-from .base import make_agg_primitive, make_trans_primitive
 from .standard import *
 from .utils import (
     get_aggregation_primitives,