+++ featuretools/tests/primitive_tests/test_feature_serialization.py
@@ -35,9 +35,9 @@ from featuretools.primitives import (
     Std,
     Sum,
     Weekday,
-    Year,
-    make_agg_primitive
+    Year
 )
+from featuretools.primitives.base import AggregationPrimitive
 from featuretools.tests.testing_utils import check_names
 
 BUCKET_NAME = "test-bucket"
@@ -85,12 +85,10 @@ def test_pickle_features(es, tmpdir):
 
 
 def test_pickle_features_with_custom_primitive(pd_es, tmpdir):
-    NewMax = make_agg_primitive(
-        lambda x: max(x),
-        name="NewMax",
-        input_types=[ColumnSchema(semantic_tags={'numeric'})],
-        return_type=ColumnSchema(semantic_tags={'numeric'}),
-        description="Calculate means ignoring nan values")
+    class NewMax(AggregationPrimitive):
+        name = "new_max"
+        input_types = [ColumnSchema(semantic_tags={'numeric'})]
+        return_type = ColumnSchema(semantic_tags={'numeric'})
 
     features_original = ft.dfs(target_dataframe_name='sessions', entityset=pd_es,
                                agg_primitives=["Last", "Mean", NewMax], features_only=True)